+++ src/main/java/com/portfoliomaker/util/StringUtil.java
@@ -4,7 +4,7 @@ import java.text.NumberFormat;
 import java.util.Locale;
 
 public class StringUtil {
-    public static String formatMoney(long money) {
+    public static String formatMoney(double money) {
         return NumberFormat.getNumberInstance(Locale.US).format(money);
     }
